requirements.txt
Django==1.6.2
argparse==1.2.1
distribute==0.6.24
django-bootstrap-toolkit==2.15.0
django-bootstrap3==4.2.0
django-debug-toolbar==1.1
django-grappelli==2.5.3
django-registration==1.0
django-registration-defaults==0.4.4
django-widget-tweaks==1.3
python-nmap==0.3.3
sqlparse==0.1.11
wsgiref==0.1.2
